WORKING_COMMIT. tokenHardCodedFetch_TestPage
If you goto: <app-base-url>/pattern
Then to: <app-base-url>/tokenHardCodedFetch_TestPage
The web-app will fetch the correct token using the hash value as a key.

It is a bit hackish because of the ordering. Hence the instructions
in the commit. The next thing to do is to refactor our token_table
So that it contains:
1. id
2. hash
3. comment
fields.
The id & hash should be marked BOTH as primary keys if possible.
Or as "must be unique" not together, but separately.
The id will be a sequential counter that counts from zero and up
by 1 for each new entry.
The HASH will be a hash of that ID that is used as our actual token.
Thus:
We have a token.id token.hash token.comment
The comment is just for debugging. Permentant part of the system.
Yes,I think debug code should be permenantly stuck in some production
code.

-John Mark Isaac Madison.
Now I am going to walk home. It is 9:37 PM. Will be home at 10:37.