Changed design philosophy of TransUtil.
We will need to test-run this app now.
I just wrote a lot of code.
If everything is okay, everything should run just as it
did before this commit.

If this commit WORKS we will make one more commit
after this notifying that the commit is in a working
state.