PREV_COMMIT_WORKS, This commit? Not sure.
The @path annotation on a class must be unique in jersy.
The @path annotation on a function must just be unique
within the scope of that class.

Even if the fully assembled path of the servlet is different
Thant the fully assembled path of another servlet,
There will still be a crash of all servlets if there are
matching @path("someWord/") on two servlets class delcarations.