WORKING_COMMIT: Servlets extend BaseServlet.java
All production servlets now extend BaseServlet.java.
So we can make new serlvets easier since we can wire up stubs
and ping them as part of the process.

There is one test servlet I made at beginning of project that
does not extend BaseServlet. It is part of the scaffolding of
working up to the final project and will eventually be scrapped.