DEPLOY_CHECK. import.sql did weird stuff again.
Database tables are no longer aligned as expected.
However, the join table for what riddles go with what
rhymes still works.

But with IDENTICAL riddles and rhymes, there are possible
Question+Answer combinations that are correct but will be
graded as WRONG.

Example: What color is the sky?
You say: BLUE
Program says: It's not BLUE, you are wrong, it is BLUE.
Because "BLUE!=BLUE" when there are two answer entries with
identical text.