Code base now forces conventions for entities.
Application will crash with error message
conventions are not followed. For how
to create an entity class.

Reasoning: strict typing in languages
makes errors less common.
Enforcing extra strictness on top of
that will prevent even more errors.

The code will also have a will to
enforce its original design.
Over time code bases become spaghetti.
Hopefully this measure of the code
looking at it's own source and
throwing errors if it is not up to
convention/quality standards will
make this code have a longer shelf life.

Or just make it a pain in the ass to
refactor. :3