SAFTEY_SAVE: Site up and running. "Test" folder...
Be careful when refactoring "test" folder.
Tried to rename it as "app". Thought all went well because
build worked. Then tried to go to welcome.jsp only to find
that NONE OF MY WEB PAGES WORKED! And there wasn't even
an error message to be found anywhere! It was kind of scary.
Especially since I need to have all this stuff ready to show
Drew tomorrow.